+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -6,6 +6,8 @@
 ([#1127](https://github.com/census-instrumentation/opencensus-python/pull/1127))
 - Fix failure counting statsbeat - refactor status code logic in transport
 ([#1132](https://github.com/census-instrumentation/opencensus-python/pull/1132))
+- Use logging handler close instead of custom atexit hook
+([#1134](https://github.com/census-instrumentation/opencensus-python/pull/1134))
 
 ## 1.1.4
 Released 2022-04-20